infrastructure/training-hf/requirements.txt
accelerate>=1.1.0
bitsandbytes>=0.45.0
datasets>=3.1.0
peft>=0.13.0
PyYAML>=6.0.2
safetensors>=0.4.5
transformers>=4.46.0
